# apps/gateway/spec/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Play-Money Poker MVP API
  version: 0.1.0
servers:
  - url: http://localhost:4000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Optional idempotency key for safe retries of mutating requests.
      schema:
        type: string
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
...
    TableConfig:
      type: object
      required: [smallBlind, bigBlind, maxPlayers, startingStack, bettingStructure]
      properties:
        smallBlind:
          type: integer
        bigBlind:
          type: integer
        ante:
          type: integer
          nullable: true
        maxPlayers:
          type: integer
        startingStack:
          type: integer
        bettingStructure:
          type: string
          enum: [NoLimit]
...
    TableSeat:
      type: object
      required: [seatId, userId, stack, status]
      properties:
        seatId:
          type: integer
        userId:
          type: string
          nullable: true
        stack:
          type: integer
        status:
          type: string
          enum: [empty, active, folded, all_in, disconnected]
    SpectatorView:
      type: object
      required: [userId, status]
      properties:
        userId:
          type: string
        nickname:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, disconnected]
    TableState:
      type: object
      required: [tableId, name, ownerId, config, seats, spectators, status, hand, version]
      properties:
        tableId:
          type: string
        name:
          type: string
        ownerId:
          type: string
        config:
          $ref: '#/components/schemas/TableConfig'
        seats:
          type: array
          items:
            $ref: '#/components/schemas/TableSeat'
        spectators:
          type: array
          items:
            $ref: '#/components/schemas/SpectatorView'
        status:
          type: string
          enum: [lobby, in_hand]
        hand:
          allOf:
            - $ref: '#/components/schemas/HandState'
          nullable: true
        version:
          type: integer
...
    ModerationRequest:
      type: object
      required: [seatId]
      properties:
        seatId:
          type: integer
          minimum: 0
          maximum: 8
...
    Pot:
      type: object
      required: [amount, eligibleSeatIds]
      properties:
        amount:
          type: integer
        eligibleSeatIds:
          type: array
          items:
            type: integer
    HandState:
      type: object
      required:
        [
          handId,
          tableId,
          buttonSeat,
          smallBlindSeat,
          bigBlindSeat,
          communityCards,
          pots,
          currentStreet,
          currentTurnSeat,
          currentBet,
          minRaise,
          raiseCapped,
          roundContributions,
          totalContributions,
          actedSeats,
          actionTimerDeadline,
          startedAt,
          endedAt,
          bigBlind
        ]
      properties:
        handId:
          type: string
        tableId:
          type: string
        buttonSeat:
          type: integer
        smallBlindSeat:
          type: integer
        bigBlindSeat:
          type: integer
        communityCards:
          type: array
          items:
            type: string
        pots:
          type: array
          items:
            $ref: '#/components/schemas/Pot'
        currentStreet:
          type: string
          enum: [preflop, flop, turn, river, showdown, ended]
        currentTurnSeat:
          type: integer
        currentBet:
          type: integer
        minRaise:
          type: integer
        raiseCapped:
          type: boolean
        roundContributions:
          type: object
          additionalProperties:
            type: integer
        totalContributions:
          type: object
          additionalProperties:
            type: integer
        actedSeats:
          type: array
          items:
            type: integer
        actionTimerDeadline:
          type: string
          nullable: true
        startedAt:
          type: string
        endedAt:
          type: string
          nullable: true
        bigBlind:
          type: integer
        winners:
          type: array
          items:
            type: integer
        holeCards:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        deck:
          type: array
          items:
            type: string
...
security:
  - bearerAuth: []
paths:
...
  /api/tables/{tableId}/moderation/kick:
    post:
      summary: Kick a player (owner only)
      parameters:
        - name: tableId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModerationRequest'
      responses:
        '200':
          description: Player kicked
          content:
            application/json:
              schema:
                type: object
                required: [tableId, seatId, userId, action, tableState]
                properties:
                  tableId:
                    type: string
                  seatId:
                    type: integer
                  userId:
                    type: string
                  action:
                    type: string
                    enum: [kick]
                  tableState:
                    $ref: '#/components/schemas/TableState'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Table or seat not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Unable to kick
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tables/{tableId}/moderation/mute:
    post:
      summary: Mute a player (owner only)
      parameters:
        - name: tableId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModerationRequest'
      responses:
        '200':
          description: Player muted
          content:
            application/json:
              schema:
                type: object
                required: [tableId, seatId, userId, action]
                properties:
                  tableId:
                    type: string
                  seatId:
                    type: integer
                  userId:
                    type: string
                  action:
                    type: string
                    enum: [mute]
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Table or seat not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...